# docs/swagger.yaml
basePath: /api
definitions:
  domain.Author:
    properties:
      avatar_emoji:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
  domain.CreatePostRequest:
    properties:
      body:
        example: I think this stock is going up!
        type: string
      ticker:
        example: AAPL
        type: string
    type: object
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.PaginatedResponse:
    properties:
      data: {}
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  domain.PostResponse:
    properties:
      author:
        $ref: '#/definitions/domain.Author'
      body:
        type: string
      comments_count:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_liked:
        type: boolean
      likes_count:
        type: integer
      ticker:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for Trading Chat application with user authentication and
    management.
  title: Trading Chat API
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Get all posts, paginated
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated posts
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get global feed
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created post
          schema:
            $ref: '#/definitions/domain.PostResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      description: Delete a post (only owner can delete)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - Posts
    get:
      description: Get a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post details
          schema:
            $ref: '#/definitions/domain.PostResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a single post
      tags:
      - Posts
...
  /posts/following:
    get:
      description: Get posts from followed users, paginated
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated posts
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get following feed
      tags:
      - Posts
  /posts/user/{id}:
    get:
      description: Get posts by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated posts
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's posts
      tags:
      - Posts
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"